requirements.txt
pydantic
requests
fastapi
pymongo
python-dotenv
parsel